fix: change code for the function modules/cjsToEsm.js
