+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+  "prettier.singleQuote": true,
+  "prettier.withNodeModules": true
+}
